fix(ANSH-218): seperate modules
