+++ src/photographer/photographer.service.ts
@@ -7,21 +7,17 @@ import {
 } from '@prisma/client';
 import { VisibilityDto } from './dto/visibility.dto';
 import {
-  HttpException,
-  HttpStatus,
   Injectable,
   Logger,
   NotFoundException,
 } from '@nestjs/common';
-import { Photographer, User } from '@prisma/client';
+import { Photographer } from '@prisma/client';
 import { contactDetailsDto } from './dto/contactDetails.dto';
 import { FeedDto } from './dto/feed.dto';
 import { FeedLikeDto } from './dto/feedLike.dto';
 import { FeedSaveDto } from './dto/feedSave.dto';
 import { DeleteFeedDto } from './dto/deleteFeed.dto';
 import { CaptionDto } from './dto/caption.dto';
-import { bankDetailsDto } from './dto/bankDetails.dto';
-import { AlbumImagesDto, AlbumsDto, UpdateAlbumCoverDto, updateAlbumDto } from './dto/album.dto';
 import { updatePackageDto } from './dto/updatePackage.dto';
 import { createPackageDto } from './dto/createPackage.dto';
 import { deletePackageDto } from './dto/deletePackage.dto';
@@ -29,10 +25,7 @@ import { deletePackageDto } from './dto/deletePackage.dto';
 import { createEventDto } from './dto/createEvent.dto';
 import { updateEventDto } from './dto/updateEvent.dto';
 import { deleteEventDto } from './dto/deleteEvent.dto';
-import { NotifyService } from '../notification/notify.service';
-import { CreateNotifyDto } from 'src/notification/dto/create-notify.dto';
 import { ClientBookingDto } from './dto/clientBooking.dto';
-import { EarningsDto } from './dto/earnings.dto';
 
 @Injectable()
 export class PhotographerService {
@@ -381,495 +374,6 @@ export class PhotographerService {
     await Promise.all(updatePromises);
   }
 
-  // ------- settings services ---------
-  async getBankDetails(photographerId: string) {
-    this.logger.log(
-      `Attempting to fetch bank details for photographer with ID: ${photographerId}`,
-    );
-
-    try {
-      // Fetching the bank details associated with the specified photographer ID
-      const bankDetails = await this.prisma.bankDetails.findUnique({
-        where: {
-          photographerId: photographerId,
-        },
-        select: {
-          id: true,
-          bankName: true,
-          accountNumber: true,
-          accountName: true,
-          accountBranch: true,
-          accountBranchCode: true,
-        },
-      });
-
-      // Handling the case where bank details are not found
-      this.logger.log(
-        `Successfully fetched category for photographer with ID: ${photographerId}`,
-      );
-      return bankDetails;
-    } catch (error) {
-      this.logger.error(
-        `Failed to fetch category for photographer with ID: ${photographerId}`,
-        error.stack,
-      );
-      if (error instanceof NotFoundException) {
-        throw error; // Re-throw if it's a not found exception.
-      }
-      throw new HttpException(
-        'Error fetching category',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async updateBankDetails(userId: string, dto: bankDetailsDto) {
-    const excistingBankDetails = await this.prisma.bankDetails.findUnique({
-      where: { photographerId: userId },
-    });
-
-    if (excistingBankDetails) {
-      await this.prisma.bankDetails.update({
-        where: { photographerId: userId },
-        data: {
-          bankName: dto.bankName,
-          accountNumber: dto.accountNumber,
-          accountName: dto.accountName,
-          accountBranch: dto.accountBranch,
-          accountBranchCode: dto.accountBranchCode
-            ? dto.accountBranchCode
-            : null,
-        },
-      });
-    } else {
-      await this.prisma.bankDetails.create({
-        data: {
-          photographer: {
-            connect: {
-              userId: userId,
-            },
-          },
-          bankName: dto.bankName,
-          accountNumber: dto.accountNumber,
-          accountName: dto.accountName,
-          accountBranch: dto.accountBranch,
-          accountBranchCode: dto.accountBranchCode
-            ? dto.accountBranchCode
-            : undefined,
-        },
-      });
-    }
-    return await this.prisma.bankDetails.findUnique({
-      where: { photographerId: userId },
-    });
-  }
-
-  async getAllCategories() {
-    this.logger.log('Attempting to fetch all photographer categories');
-
-    try {
-      // Fetching the predefined categories from the constants
-      const categories = PhotographerCategory;
-
-      this.logger.log('Successfully fetched all photographer categories');
-      return categories;
-    } catch (error) {
-      this.logger.error('Failed to fetch photographer categories', error.stack);
-
-      // Throwing an HTTP exception with internal server error status
-      throw new HttpException(
-        'Error fetching categories',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async getCategoryById(id: string) {
-    this.logger.log(
-      `Attempting to fetch category for photographer with ID: ${id}`,
-    );
-
-    try {
-      const category = await this.prisma.photographer.findUnique({
-        where: {
-          userId: id,
-        },
-        select: {
-          category: true,
-        },
-      });
-      
-      this.logger.log(
-        `Successfully fetched category for photographer with ID: ${id}`,
-      );
-      return category;
-    } catch (error) {
-      this.logger.error(
-        `Failed to fetch category for photographer with ID: ${id}`,
-        error.stack,
-      );
-      if (error instanceof NotFoundException) {
-        throw error; // Re-throw if it's a not found exception.
-      }
-      throw new HttpException(
-        'Error fetching category',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async updateCategory(userId: string, data: Partial<Photographer>) {
-    this.logger.log(
-      `Attempting to update category for photographer with ID: ${userId}`,
-    );
-
-    try {
-      const updatedPhotographer = await this.prisma.photographer.update({
-        where: {
-          userId: userId,
-        },
-        include: {
-          user: true,
-        },
-        data: {
-          category: {
-            set: data.category,
-          },
-        },
-      });
-
-      this.logger.log(
-        `Successfully updated category for photographer with ID: ${userId}`,
-      );
-      return updatedPhotographer;
-    } catch (error) {
-      this.logger.error(
-        `Failed to update category for photographer with ID: ${userId}`,
-        error.stack,
-      );
-
-      // Handling case where the photographer is not found
-      if (error.code === 'P2025') {
-        // P2025 is Prisma's code for a record not found error
-        throw new NotFoundException(
-          `Photographer with ID: ${userId} not found`,
-        );
-      }
-
-      // General error handling
-      throw new HttpException(
-        'Error updating category',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  // ------- album services ---------
-
-  async createAlbum(dto: AlbumsDto) {
-    this.logger.log(
-      `Attempting to create a new album for photographer with ID: ${dto.photographerId}`,
-    );
-
-    try {
-      // Creating a new album in the database with the provided details
-      const newAlbum = await this.prisma.album.create({
-        data: {
-          name: dto.name,
-          description: dto.description,
-          visibility: dto.visibility,
-          price: dto.price,
-          photographer: {
-            connect: {
-              userId: dto.photographerId,
-            },
-          },
-        },
-        include: {
-          images: true, // Includes related images in the response
-        },
-      });
-
-      this.logger.log(
-        `Successfully created album '${dto.name}' for photographer with ID: ${dto.photographerId}`,
-      );
-      return newAlbum;
-    } catch (error) {
-      this.logger.error(
-        `Failed to create album '${dto.name}' for photographer with ID: ${dto.photographerId}`,
-        error.stack,
-      );
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error creating album',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async editAlbum(dto: updateAlbumDto) {
-    try {
-      // Updating the album in the database with the provided details
-      const updatedAlbum = await this.prisma.album.update({
-        where: {
-          id: dto.albumId,
-        },
-        data: {
-          name: dto.name,
-          description: dto.description,
-          visibility: dto.visibility,
-          price: dto.price,
-        },
-      });
-
-      this.logger.log(`Successfully updated album with ID: ${dto.albumId}`);
-      return updatedAlbum;
-    } catch (error) {
-      this.logger.error(
-        `Failed to update album with ID: ${dto.albumId}`,
-        error.stack,
-      );
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error updating album',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async getAlbums(id: string) {
-    this.logger.log(
-      `Attempting to fetch all albums for photographer with ID: ${id}`,
-    );
-
-    try {
-      // Fetching all albums for the specified photographer ID, including associated images
-      const albums = await this.prisma.album.findMany({
-        where: {
-          photographerId: id,
-        },
-        include: {
-          images: true, // Includes related images in the response
-        },
-      });
-
-      if (albums.length === 0) {
-        this.logger.warn(`No albums found for photographer with ID: ${id}`);
-        // Optionally: You can return an empty array or a specific message if no albums are found
-        return [];
-      }
-
-      this.logger.log(
-        `Successfully fetched ${albums.length} albums for photographer with ID: ${id}`,
-      );
-      return albums;
-    } catch (error) {
-      this.logger.error(
-        `Failed to fetch albums for photographer with ID: ${id}`,
-        error.stack,
-      );
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error fetching albums',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async getAlbum(id: string) {
-    this.logger.log(`Attempting to fetch album with ID: ${id}`);
-
-    try {
-      // Fetching the album from the database by its ID
-      const album = await this.prisma.album.findMany({
-        where: {
-          id: id,
-        },
-      });
-
-      if (album.length === 0) {
-        this.logger.warn(`No album found with ID: ${id}`);
-        // Throwing a 404 exception if no album is found
-        throw new HttpException('Album not found', HttpStatus.NOT_FOUND);
-      }
-
-      this.logger.log(`Successfully fetched album with ID: ${id}`);
-      return album;
-    } catch (error) {
-      this.logger.error(`Failed to fetch album with ID: ${id}`, error.stack);
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error fetching album',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async getImages(id: string) {
-    this.logger.log(`Attempting to fetch images for album with ID: ${id}`);
-
-    try {
-      // Fetching all images associated with the specified album ID
-      const images = await this.prisma.albumImage.findMany({
-        where: {
-          albumId: id,
-        },
-      });
-
-      if (images.length === 0) {
-        this.logger.warn(`No images found for album with ID: ${id}`);
-      } else {
-        this.logger.log(
-          `Successfully fetched ${images.length} images for album with ID: ${id}`,
-        );
-      }
-
-      return images;
-    } catch (error) {
-      this.logger.error(
-        `Failed to fetch images for album with ID: ${id}`,
-        error.stack,
-      );
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error fetching images',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async deleteAlbum(id: string) {
-    this.logger.log(`Attempting to delete album with ID: ${id}`);
-
-    try {
-      // Deleting the album with the specified ID
-      const deletedAlbum = await this.prisma.album.delete({
-        where: {
-          id: id,
-        },
-      });
-
-      this.logger.log(`Successfully deleted album with ID: ${id}`);
-      return deletedAlbum;
-    } catch (error) {
-      this.logger.error(`Failed to delete album with ID: ${id}`, error.stack);
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error deleting album',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async addImages(dto: AlbumImagesDto) {
-    this.logger.log(
-      `Attempting to add images to album with ID: ${dto.albumId}`,
-    );
-
-    try {
-      // Checking if the album exists
-      const album = await this.prisma.album.findUnique({
-        where: {
-          id: dto.albumId,
-        },
-      });
-
-      if (!album) {
-        this.logger.warn(`Album not found with ID: ${dto.albumId}`);
-        throw new NotFoundException('Album not found');
-      }
-
-      // Adding the images to the album
-      const images = await Promise.all(
-        dto.images.map((imageUrl) =>
-          this.prisma.albumImage.create({
-            data: {
-              image: imageUrl,
-              album: {
-                connect: {
-                  id: dto.albumId,
-                },
-              },
-            },
-          }),
-        ),
-      );
-
-      this.logger.log(
-        `Successfully added ${dto.images.length} images to album with ID: ${dto.albumId}`,
-      );
-      return images;
-    } catch (error) {
-      this.logger.error(
-        `Failed to add images to album with ID: ${dto.albumId}`,
-        error.stack,
-      );
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error adding images to album',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async deleteImage(id: string) {
-    this.logger.log(`Attempting to delete image with ID: ${id}`);
-
-    try {
-      // Deleting the image with the specified ID
-      const deletedImage = await this.prisma.albumImage.delete({
-        where: {
-          id: id,
-        },
-      });
-
-      this.logger.log(`Successfully deleted image with ID: ${id}`);
-      return deletedImage;
-    } catch (error) {
-      this.logger.error(`Failed to delete image with ID: ${id}`, error.stack);
-
-      // Throwing a generic HTTP exception with internal server error status
-      throw new HttpException(
-        'Error deleting image',
-        HttpStatus.INTERNAL_SERVER_ERROR,
-      );
-    }
-  }
-
-  async updateAlbumCover(dto:UpdateAlbumCoverDto){
-    this.logger.log(`Attempting to update cover image for album ID: ${dto.albumId}`);
-
-    try {
-      // Check if the album exists
-      const album = await this.prisma.album.findUnique({
-        where: { id: dto.albumId },
-      });
-
-      if (!album) {
-        this.logger.warn(`Album with ID ${dto.albumId} not found`);
-        throw new NotFoundException(`Album with ID ${dto.albumId} not found`);
-      }
-
-      // Update the album's cover image
-      const updatedAlbum = await this.prisma.album.update({
-        where: { id: dto.albumId },
-        data: { coverImage: dto.coverImage },
-      });
-
-      this.logger.log(`Cover image updated successfully for album ID: ${dto.albumId}`);
-      return updatedAlbum;
-    } catch (error) {
-      this.logger.error(`Error updating cover image for album ID: ${dto.albumId}`, error.stack);
-      throw error; // Re-throw the error to be handled by the controller
-    }
-  }
-
   // ------- feed services ---------
 
   async getFeed(id: string) {
@@ -1182,81 +686,6 @@ export class PhotographerService {
         },
       },
     });
-  }
-
-  // ------- payment services ---------
-
-  async getPayments(id: string) {
-    this.logger.log(`Fetching payments for photographer with ID: ${id}`);
-    try {
-      const payments = await this.prisma.payment.findMany({
-        where: {
-          photographerId: id,
-        },
-        include: {
-          client: {
-            select: {
-              name: true,
-            },
-          },
-        },
-      });
-      this.logger.log(
-        `Successfully fetched payments for photographer with ID: ${id}`,
-      );
-      return payments;
-    } catch (error) {
-      this.logger.error(
-        `Failed to fetch payments for photographer with ID: ${id}`,
-        error.stack,
-      );
-      throw new Error('Error fetching payments');
-    }
-  }
-
-  async getEarnings(id: string) {
-    this.logger.log(`Calculating earnings for photographer with ID: ${id}`);
-    try {
-      // Fetch payments for the given photographer
-      const payments = await this.prisma.payment.findMany({
-        where: {
-          photographerId: id,
-        },
-        select: {
-          amount: true,
-          status: true,
-        },
-      });
-
-      let paidTot = 0;
-      let pendingTot = 0;
-
-      // Calculate total paid and pending amounts
-      payments.forEach((payment) => {
-        if (payment.status === 'PAID') {
-          paidTot += payment.amount;
-        } else if (payment.status === 'PENDING') {
-          pendingTot += payment.amount;
-        }
-      });
-
-      // Create the earnings DTO
-      const earningsDtoData = new EarningsDto();
-      earningsDtoData.fees = 0.1 * paidTot; // 10% fees
-      earningsDtoData.totalAmount = paidTot - earningsDtoData.fees; // Net earnings after fees
-      earningsDtoData.pending = pendingTot; // Total pending payments
-
-      this.logger.log(
-        `Successfully calculated earnings for photographer with ID: ${id}`,
-      );
-      return earningsDtoData;
-    } catch (error) {
-      this.logger.error(
-        `Failed to calculate earnings for photographer with ID: ${id}`,
-        error.stack,
-      );
-      throw new Error('Error calculating earnings');
-    }
   }
 
   //------- booking services ---------
